+++ src/styles/main_window/__init__.py
@@ -1,13 +1,14 @@
-from .shell_styles import (
-    LauncherShellPalette,
-    PALETTE,
-    get_launcher_shell_stylesheet,
-    resolve_launcher_asset,
-)
+from styles.compose import build_main_window_qss, get_main_window_stylesheet
+from styles.shell.launcher_shell import LauncherShellPalette, PALETTE, get_launcher_shell_stylesheet, resolve_launcher_asset
+from styles.theme import THEME, get_theme
 
 __all__ = [
+    "THEME",
     "LauncherShellPalette",
     "PALETTE",
+    "build_main_window_qss",
+    "get_main_window_stylesheet",
+    "get_theme",
     "get_launcher_shell_stylesheet",
     "resolve_launcher_asset",
 ]
