+++ src/styles/main_window/base.py
@@ -1,0 +1,3 @@
+from styles.base import BASE_QSS
+
+__all__ = ["BASE_QSS"]
