+++ src/styles/main_window/chat.py
@@ -1,0 +1,3 @@
+from styles.pages.sandbox import CHAT_SCROLL_QSS, CHAT_WIDGETS_QSS, SANDBOX_PAGE_QSS
+
+__all__ = ["CHAT_SCROLL_QSS", "CHAT_WIDGETS_QSS", "SANDBOX_PAGE_QSS"]
