+++ src/styles/main_window/components.py
@@ -1,0 +1,3 @@
+from styles.components import COMPONENTS_QSS
+
+__all__ = ["COMPONENTS_QSS"]
