+++ src/styles/main_window/compose.py
@@ -1,0 +1,3 @@
+from styles.compose import build_main_window_qss, get_main_window_stylesheet
+
+__all__ = ["build_main_window_qss", "get_main_window_stylesheet"]
