+++ src/styles/main_window/home_page.py
@@ -1,0 +1,3 @@
+from styles.pages.home import HOME_PAGE_QSS
+
+__all__ = ["HOME_PAGE_QSS"]
