+++ src/styles/main_window/settings.py
@@ -1,0 +1,3 @@
+from styles.pages.settings import SETTINGS_PAGE_QSS as SETTINGS_QSS
+
+__all__ = ["SETTINGS_QSS"]
