+++ src/styles/main_window/theme.py
@@ -1,0 +1,3 @@
+from styles.theme import THEME, get_theme
+
+__all__ = ["THEME", "get_theme"]
