+++ src/styles/main_window/window_shell.py
@@ -1,0 +1,3 @@
+from styles.shell.main_window import MAIN_WINDOW_SHELL_QSS as SHELL_QSS
+
+__all__ = ["SHELL_QSS"]
