+++ src/styles/pages/__init__.py
@@ -1,0 +1,15 @@
+from .home import HOME_PAGE_QSS
+from .logs import LOGS_PAGE_QSS
+from .news import NEWS_PAGE_QSS
+from .sandbox import CHAT_SCROLL_QSS, CHAT_WIDGETS_QSS, SANDBOX_PAGE_QSS
+from .settings import SETTINGS_PAGE_QSS
+
+__all__ = [
+    "CHAT_SCROLL_QSS",
+    "CHAT_WIDGETS_QSS",
+    "HOME_PAGE_QSS",
+    "LOGS_PAGE_QSS",
+    "NEWS_PAGE_QSS",
+    "SANDBOX_PAGE_QSS",
+    "SETTINGS_PAGE_QSS",
+]
