+++ src/styles/pages/logs.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+LOGS_PAGE_QSS = ""
