+++ src/styles/pages/news.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+NEWS_PAGE_QSS = ""
