+++ src/styles/shell/__init__.py
@@ -1,0 +1,10 @@
+from .launcher_shell import LauncherShellPalette, PALETTE, get_launcher_shell_stylesheet, resolve_launcher_asset
+from .main_window import MAIN_WINDOW_SHELL_QSS
+
+__all__ = [
+    "LauncherShellPalette",
+    "MAIN_WINDOW_SHELL_QSS",
+    "PALETTE",
+    "get_launcher_shell_stylesheet",
+    "resolve_launcher_asset",
+]
